refactor: split EmailViewer into Header, EmailList, EmailDetails
