+++ packages/mochi/src/templates/EmailViewer/EmailViewer.svelte
@@ -1,72 +1,11 @@
 <script lang="ts">
-  import EmailView from './EmailView.svelte';
-  import MsgItem from './MsgItem.svelte';
+  import EmailDetails from './EmailDetails.svelte';
+  import EmailList from './EmailList.svelte';
+  import Header from './Header.svelte';
   import OutboxSync from './OutboxSync.svelte';
-
-  interface StoredAttachment {
-    filename: string;
-    contentType?: string;
-    size: number;
-  }
-  interface StoredEmail {
-    id: string;
-    sentAt: number;
-    from: string;
-    to: string[];
-    cc?: string[];
-    bcc?: string[];
-    replyTo?: string;
-    subject: string;
-    html?: string;
-    text?: string;
-    headers?: Record<string, string>;
-    attachments?: StoredAttachment[];
-  }
-  interface EmailListItem {
-    id: string;
-    sentAt: number;
-    from: string;
-    to: string[];
-    subject: string;
-    hasHtml: boolean;
-    hasText: boolean;
-    attachmentCount: number;
-  }
+  import type { EmailListItem, StoredEmail } from './types';
 
   let { emails, selected, basePath }: { emails: EmailListItem[]; selected: StoredEmail | null; basePath: string } = $props();
-
-  const example = `Mochi.email({ to: 'you@app.test', subject: 'Hello', html: '<b>Hi!</b>' })`;
-
-  function fmtDate(ts: number): string {
-    return new Date(ts).toLocaleString();
-  }
-
-  function fmtSize(bytes: number): string {
-    if (bytes < 1024) {
-      return `${bytes} B`;
-    }
-    if (bytes < 1024 * 1024) {
-      return `${(bytes / 1024).toFixed(1)} KB`;
-    }
-    return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-  }
-
-  const metaRows = $derived(
-    selected
-      ? (
-          [
-            { label: 'From', value: selected.from },
-            { label: 'To', value: selected.to.join(', ') },
-            selected.cc?.length ? { label: 'Cc', value: selected.cc.join(', ') } : null,
-            selected.bcc?.length ? { label: 'Bcc', value: selected.bcc.join(', ') } : null,
-            selected.replyTo ? { label: 'Reply-To', value: selected.replyTo } : null,
-            { label: 'Date', value: fmtDate(selected.sentAt) },
-          ] as Array<{ label: string; value: string } | null>
-        ).filter((r): r is { label: string; value: string } => r !== null)
-      : [],
-  );
-
-  const headerEntries = $derived(selected?.headers ? Object.entries(selected.headers) : []);
 </script>
 
 <svelte:head>
@@ -77,97 +16,11 @@
 <OutboxSync viewedId={selected?.id ?? null} allIds={emails.map((e) => e.id)} mochi:hydrate />
 
 <div class="shell">
-  <header class="topbar">
-    <div class="brand">
-      <span class="mark">🍡</span>
-      <span class="wordmark">mochi <em>outbox</em></span>
-      <span class="dev-pill">dev</span>
-    </div>
-    <div class="topbar-actions">
-      <span class="count">{emails.length} captured</span>
-      {#if emails.length > 0}
-        <form method="POST" action="?/clear" class="clear-form">
-          <button type="submit" name="clear" class="clear-btn">Clear</button>
-        </form>
-      {/if}
-    </div>
-  </header>
+  <Header count={emails.length} />
 
   <div class="split">
-    <aside class="list-pane">
-      {#if emails.length === 0}
-        <div class="list-empty">Empty</div>
-      {:else}
-        <ul class="msg-list">
-          {#each emails as e (e.id)}
-            <li>
-              <MsgItem
-                id={e.id}
-                subject={e.subject}
-                to={e.to}
-                sentAt={e.sentAt}
-                hasHtml={e.hasHtml}
-                hasText={e.hasText}
-                attachmentCount={e.attachmentCount}
-                active={selected?.id === e.id}
-                {basePath}
-                mochi:hydrate
-              />
-            </li>
-          {/each}
-        </ul>
-      {/if}
-    </aside>
-
-    <main class="detail-pane">
-      {#if !selected}
-        <div class="detail-empty">
-          <div class="empty-mark">✉️</div>
-          <h2>No mail captured yet</h2>
-          <p>Messages sent through the <code>dev</code> transport land here. Send one from your app:</p>
-          <pre class="empty-code"><code>{example}</code></pre>
-        </div>
-      {:else}
-        <div class="detail-head">
-          <h1 class="detail-subject">{selected.subject || '(no subject)'}</h1>
-          <dl class="meta-grid">
-            {#each metaRows as row (row.label)}
-              <div class="meta-item">
-                <dt>{row.label}</dt>
-                <dd>{row.value}</dd>
-              </div>
-            {/each}
-          </dl>
-
-          {#if headerEntries.length > 0}
-            <details class="headers">
-              <summary>Custom headers ({headerEntries.length})</summary>
-              <dl class="header-list">
-                {#each headerEntries as [k, v] (k)}
-                  <div class="header-item">
-                    <dt>{k}</dt>
-                    <dd>{v}</dd>
-                  </div>
-                {/each}
-              </dl>
-            </details>
-          {/if}
-        </div>
-
-        {#if selected.attachments && selected.attachments.length > 0}
-          <div class="attachments">
-            {#each selected.attachments as a (a.filename)}
-              <span class="attach-chip">
-                <span class="attach-name">{a.filename}</span>
-                <span class="attach-size">{fmtSize(a.size)}</span>
-              </span>
-            {/each}
-          </div>
-        {/if}
-
-        <EmailView mochi:hydrate html={selected.html} text={selected.text} />
-      {/if}
-    </main>
+    <EmailList {emails} selectedId={selected?.id ?? null} {basePath} />
+    <EmailDetails {selected} />
   </div>
 </div>
 
@@ -255,83 +108,6 @@
     padding: 1.25rem 1.25rem 3rem;
   }
 
-  .topbar {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 1rem;
-    flex-wrap: wrap;
-    padding: 0.4rem 0.2rem 1rem;
-    border-bottom: 1px solid var(--ev-border);
-    margin-bottom: 1.25rem;
-  }
-  .brand {
-    display: inline-flex;
-    align-items: baseline;
-    gap: 0.5rem;
-  }
-  .mark {
-    font-size: 1.35rem;
-    line-height: 1;
-  }
-  .wordmark {
-    font-family: var(--ev-font-serif);
-    font-size: 1.25rem;
-    font-weight: 500;
-    color: var(--ev-text);
-    letter-spacing: -0.01em;
-  }
-  .wordmark em {
-    font-style: normal;
-    color: var(--ev-accent);
-  }
-  .dev-pill {
-    align-self: center;
-    font-size: 0.6rem;
-    font-weight: 700;
-    text-transform: uppercase;
-    letter-spacing: 0.1em;
-    color: var(--ev-accent-soft-text);
-    background: var(--ev-accent-soft);
-    padding: 0.15rem 0.45rem;
-    border-radius: 999px;
-  }
-  .topbar-actions {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.85rem;
-  }
-  .count {
-    font-size: 0.78rem;
-    color: var(--ev-text-subtle);
-    font-family: var(--ev-font-mono);
-    font-variant-numeric: tabular-nums;
-  }
-  .clear-form {
-    margin: 0;
-  }
-  .clear-btn {
-    appearance: none;
-    font-family: inherit;
-    font-size: 0.78rem;
-    font-weight: 600;
-    color: var(--ev-text-muted);
-    background: var(--ev-surface);
-    border: 1px solid var(--ev-border-strong);
-    border-radius: var(--ev-radius-sm);
-    padding: 0.35rem 0.8rem;
-    cursor: pointer;
-    transition:
-      color 0.14s ease,
-      border-color 0.14s ease,
-      background 0.14s ease;
-  }
-  .clear-btn:hover {
-    color: #7a2e20;
-    border-color: #d8a99e;
-    background: #f7e7e1;
-  }
-
   .split {
     display: grid;
     grid-template-columns: 320px 1fr;
@@ -339,188 +115,9 @@
     align-items: start;
   }
 
-  .list-pane {
-    background: var(--ev-surface);
-    border: 1px solid var(--ev-border);
-    border-radius: var(--ev-radius-lg);
-    box-shadow: var(--ev-shadow-sm);
-    overflow: hidden;
-    position: sticky;
-    top: 1rem;
-    max-height: calc(100vh - 2rem);
-    overflow-y: auto;
-  }
-  .list-empty {
-    padding: 2rem 1rem;
-    text-align: center;
-    color: var(--ev-text-subtle);
-    font-style: italic;
-    font-size: 0.85rem;
-  }
-  .msg-list {
-    list-style: none;
-    margin: 0;
-    padding: 0.4rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.15rem;
-  }
-  .detail-pane {
-    background: var(--ev-surface);
-    border: 1px solid var(--ev-border);
-    border-radius: var(--ev-radius-lg);
-    box-shadow: var(--ev-shadow-md);
-    padding: 1.5rem;
-    min-width: 0;
-  }
-
-  .detail-empty {
-    text-align: center;
-    padding: 3rem 1.5rem;
-    color: var(--ev-text-muted);
-  }
-  .empty-mark {
-    font-size: 2.5rem;
-  }
-  .detail-empty h2 {
-    font-family: var(--ev-font-serif);
-    font-weight: 500;
-    font-size: 1.25rem;
-    color: var(--ev-text);
-    margin: 0.75rem 0 0.4rem;
-  }
-  .detail-empty p {
-    font-size: 0.9rem;
-    margin: 0 auto 1rem;
-    max-width: 26rem;
-  }
-  .detail-empty code {
-    font-family: var(--ev-font-mono);
-    background: var(--ev-surface-muted);
-    padding: 0.05rem 0.3rem;
-    border-radius: 4px;
-    font-size: 0.85em;
-  }
-  .empty-code {
-    display: inline-block;
-    text-align: left;
-    margin: 0;
-    padding: 0.75rem 1rem;
-    background: var(--ev-surface-muted);
-    border: 1px solid var(--ev-border);
-    border-radius: var(--ev-radius-md);
-    font-family: var(--ev-font-mono);
-    font-size: 0.78rem;
-    color: var(--ev-text-muted);
-    overflow-x: auto;
-    max-width: 100%;
-  }
-
-  .detail-head {
-    margin-bottom: 1.25rem;
-  }
-  .detail-subject {
-    font-family: var(--ev-font-serif);
-    font-weight: 500;
-    font-size: 1.5rem;
-    letter-spacing: -0.01em;
-    color: var(--ev-text);
-    margin: 0 0 1rem;
-    word-break: break-word;
-  }
-  .meta-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(220px, 1fr));
-    gap: 0.5rem;
-    margin: 0;
-  }
-  .meta-item {
-    background: var(--ev-surface-muted);
-    border: 1px solid var(--ev-border);
-    border-radius: var(--ev-radius-md);
-    padding: 0.5rem 0.75rem;
-    min-width: 0;
-  }
-  .meta-item dt {
-    font-size: 0.65rem;
-    color: var(--ev-text-muted);
-    text-transform: uppercase;
-    letter-spacing: 0.06em;
-    font-weight: 700;
-  }
-  .meta-item dd {
-    margin: 0.15rem 0 0;
-    font-size: 0.9rem;
-    color: var(--ev-text);
-    font-weight: 500;
-    word-break: break-word;
-  }
-
-  .headers {
-    margin-top: 0.75rem;
-    font-size: 0.82rem;
-  }
-  .headers summary {
-    cursor: pointer;
-    color: var(--ev-text-muted);
-    font-weight: 600;
-  }
-  .header-list {
-    margin: 0.5rem 0 0;
-    display: flex;
-    flex-direction: column;
-    gap: 0.2rem;
-  }
-  .header-item {
-    display: flex;
-    gap: 0.6rem;
-    font-family: var(--ev-font-mono);
-    font-size: 0.76rem;
-  }
-  .header-item dt {
-    color: var(--ev-text-subtle);
-    min-width: 8rem;
-    flex-shrink: 0;
-  }
-  .header-item dd {
-    margin: 0;
-    color: var(--ev-text);
-    word-break: break-all;
-  }
-
-  .attachments {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 0.5rem;
-    margin-bottom: 1.1rem;
-  }
-  .attach-chip {
-    display: inline-flex;
-    align-items: baseline;
-    gap: 0.4rem;
-    padding: 0.3rem 0.65rem;
-    background: var(--ev-surface-muted);
-    border: 1px solid var(--ev-border);
-    border-radius: 999px;
-    font-size: 0.78rem;
-  }
-  .attach-name {
-    font-weight: 600;
-    color: var(--ev-text);
-  }
-  .attach-size {
-    color: var(--ev-text-subtle);
-    font-family: var(--ev-font-mono);
-    font-size: 0.7rem;
-  }
-
   @media (max-width: 760px) {
     .split {
       grid-template-columns: 1fr;
-    }
-    .list-pane {
-      position: static;
-      max-height: 340px;
     }
   }
 </style>
